Implement sandbox partial support
